CHANGED ALL AUTOS, DON'T REVERT TO THIS ONE
Made core's a constructor so now it'll error out if you don't include it.
Changed ALL of the autonomous code to be able to do simultaneous actions now. Should be good, we'll see though.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Auto/ArmAutoCore.java
@@ -46,7 +46,7 @@ public class ArmAutoCore {
         Thread.sleep(timeout);
     }
 
-    public void pvtMove(double velocity, int ticks, boolean active, int timeout, Telemetry tele) throws InterruptedException {
+    public void pvtMove(double velocity, int ticks, boolean active, long timeout, Telemetry tele) throws InterruptedException {
         pvtArm.setTargetPosition(ticks);
         pvtArm.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         pvtArm.setVelocity(velocity);
